+++ packages/mock-server/tests/test-config.js
@@ -1,4 +1,4 @@
-import path from 'path';
+import path from 'node:path';
 
 const config = {
   latency: 0,
@@ -9,7 +9,7 @@ const config = {
   host: '127.0.0.1', // 0.0.0.0 or localhost causes windows tests to fail?
 
   data: path.join(import.meta.dirname, '/data'),
-  routes: path.join(import.meta.dirname, '/dummy.routes.config.json')
+  routes: path.join(import.meta.dirname, '/dummy.routes.config.json'),
 };
 
 export default config;